refactor: DRY - fetch arns names
