+++ src/components/dashboard/subsettings/AddContact.tsx
@@ -27,6 +27,7 @@ import { ExtensionStorage } from "~utils/storage";
 import { useEffect, useState } from "react";
 import browser from "webextension-polyfill";
 import { uploadUserAvatar, getUserAvatar } from "~lib/avatar";
+import { getAllArNSNames } from "~lib/arns";
 import styled from "styled-components";
 import { useLocation } from "wouter";
 import copy from "copy-to-clipboard";
@@ -107,28 +108,6 @@ export default function AddContact() {
       [name]: value
     });
   };
-
-  async function getANTsContractTxIds(owner) {
-    const response = await fetch(
-      `https://api.arns.app/v1/wallet/${owner}/contracts?type=ant`
-    );
-    const data = await response.json();
-    return data.contractTxIds;
-  }
-
-  async function getAllArNSNames(owner) {
-    const contractTxIds = await getANTsContractTxIds(owner);
-    if (contractTxIds.length === 0) return [];
-
-    const url = [
-      "https://api.arns.app/v1/contract/bLAgYxAdX2Ry-nt6aH2ixgvJXbpsEYm28NgJgyqfs-U/records?",
-      ...contractTxIds.map((txId) => `contractTxId=${txId}`)
-    ].join("&");
-
-    const response = await fetch(url);
-    const data = await response.json();
-    return data;
-  }
 
   async function fetchArnsAddresses(ownerAddress) {
     try {
