+++ src/components/dashboard/subsettings/ContactSettings.tsx
@@ -16,6 +16,7 @@ import browser from "webextension-polyfill";
 import { Edit02, Share04, Upload01 } from "@untitled-ui/icons-react";
 import { useLocation } from "wouter";
 import { uploadUserAvatar, getUserAvatar } from "~lib/avatar";
+import { getAllArNSNames } from "~lib/arns";
 import copy from "copy-to-clipboard";
 
 export default function ContactSettings({ address }: Props) {
@@ -68,28 +69,6 @@ export default function ContactSettings({ address }: Props) {
     } catch (error) {
       console.error("Error fetching ArNS addresses:", error);
     }
-  }
-
-  async function getANTsContractTxIds(owner) {
-    const response = await fetch(
-      `https://api.arns.app/v1/wallet/${owner}/contracts?type=ant`
-    );
-    const data = await response.json();
-    return data.contractTxIds;
-  }
-
-  async function getAllArNSNames(owner) {
-    const contractTxIds = await getANTsContractTxIds(owner);
-    if (contractTxIds.length === 0) return [];
-
-    const url = [
-      "https://api.arns.app/v1/contract/bLAgYxAdX2Ry-nt6aH2ixgvJXbpsEYm28NgJgyqfs-U/records?",
-      ...contractTxIds.map((txId) => `contractTxId=${txId}`)
-    ].join("&");
-
-    const response = await fetch(url);
-    const data = await response.json();
-    return data;
   }
 
   const handleInputChange = (e) => {
